Minor refactor of data collection # Description - Moves collection to data-collector directory - Creates new helper file and exports all functions
